refactor XpansionStudyReader
